+++ src/components/Sidebar.svelte
@@ -6,6 +6,7 @@
   import { writable } from 'svelte/store'
   import { setContext, tick } from 'svelte'
   import { dragged } from '../lib/drag-drop'
+  import { main_area } from '@/lib/page'
 
   const special = {
     children: ['root'],
@@ -84,13 +85,21 @@
       on:focus={() => {
         contentElement.scrollTop = 0
       }}
+      on:keydown={(e) => {
+        if (e.key === 'Escape') {
+          main_area.focus()
+        }
+      }}
     />
     <!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
     <nav
       class="items"
       tabindex="-1"
       on:keydown={(e) => {
-        if (e.key == 'Home' || e.key == 'End' || e.key == 'PageUp' || e.key == 'PageDown') {
+        if (e.key === 'Escape') {
+          e.preventDefault()
+          main_area.focus()
+        } else if (e.key == 'Home' || e.key == 'End' || e.key == 'PageUp' || e.key == 'PageDown') {
           e.preventDefault()
         } else {
           $itemHandle?.handleKey(e)