+++ src/lib/page.ts
@@ -5,6 +5,13 @@ import { newSelection } from './selection'
 
 export const scrollToIndex = writable(null as number | null)
 
+let main_area_el: HTMLElement | undefined
+export const main_area = {
+  focus() {
+    main_area_el?.focus()
+  },
+}
+
 export const selection = newSelection({
   getItemCount: () => get(page).length,
   scrollToItem: (i) => scrollToIndex.set(i),